backend/requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-marshmallow==1.2.1
marshmallow-sqlalchemy==1.1.0
flask-cors==4.0.1
gunicorn==22.0.0
werkzeug==3.0.3
flask-migrate==4.0.7
alembic==1.13.2
